+++ src/services/users/index.ts
@@ -1,4 +1,4 @@
-import { fetchAPI } from '@/utils/api';
+import { fetchAPI } from '@/services/api';
 import { FetchResult } from '../type';
 import { Profile, ResIsUsernameAvailable } from './type';
